+++ app/commands/draw.js
@@ -1,5 +1,7 @@
 import { COMMAND_DRAW } from '../../constants/command.js';
+import { SETTING_IMAGE_GENERATION_SIZE } from '../../constants/setting.js';
 import { PARTICIPANT_AI, PARTICIPANT_HUMAN } from '../../services/openai.js';
+import storage from '../../storage/index.js';
 import generateImage from '../../utils/generate-image.js';
 import Event from '../event.js';
 import { getSession, setSession } from '../sessions.js';
@@ -21,8 +23,9 @@ const execDrawCommand = async (event) => {
     .write(`${event.text}？`)
     .write(`\n${PARTICIPANT_AI}: `);
   try {
-    const { url } = await generateImage({ prompt: event.text });
-    setSession(event.userId, session.write('OK'));
+    const size = await storage.getItem(SETTING_IMAGE_GENERATION_SIZE);
+    const { url } = await generateImage({ prompt: event.text, size });
+    setSession(event.userId, session.write('OK!'));
     event.sendImage(url);
   } catch (err) {
     event.sendText(err.message);